ai-microservice/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
